Refactor dconf dump
Dumping everything into a single file was cumbersome. It makes it harder
to select specific parts to load. This format is much better in this
regard but I'm not sure it's optimal either. In fact, this new format
made it easy to discard a whole lot of content that I don't use anymore.

To load the dconf database,
one has to somehow iterate over the various files and use their path as
the dir argument for dconf. For example

    dconf load /apps/guake/ < apps/guake

Thankfully, this is easy to do in a one-liner for loop.

I'm wondering whether dumping everything into a single file was good
enough if I would have added/found some script to make it easy to select
specific values. I will keep considering and revisit it in the future.