sway: Use pacwall background
The path for the background image is the one my timer sets. Also set the
terminals to be transparent so that the image is always at least
slightly visible. It's kinda cheesy in an "OMG LEET HAXORZ" way, but boy
does it look nice, especially in combination with the gaps.