Add .stowrc
Specify --no-folding because the default behavior plays terribly with
our setup. Some programs generate lots of files in their directories
which we don't want. If stow links to a directory instead of just the
files we want, we end up polluting our git repo with lots of useless
files. On the flip side, I haven't come across a scenario where
no-folding gives a worse result.