Workaround for broken ssh sessions
If we use kitty or any other uncommon terminal to ssh into a system, the
terminfo is likely to be missing so we get a broken terminal. Setting
this to the common xterm fixes this, though I guess it reduces the
features that we can use. Maybe there is a better workaround or a more
appropriate terminfo. I will have to come back to this once I understand
it more.