sway: Fix volume going above 100%
In fact it pretty much went to infinity. Changing the program to amixer
automatically fixes this as it enforces limits anyway. Maybe at some
point I want to investigate solutions that enforce a limit that is >100%
to help with sources that have too low audio, but I'm fine with this for
now.