zsh: Change TERM for ssh
This should be more appropriate than the old one, but I'm still not sure
if it's the best choice. At some point I'll have to delve into the issue
and come up with a terminfo that is close to kitty but likely to be
present in most systems I ssh into.