zsh: Create alias for kitty's icat
This utility can display image directly in the terminal which is pretty
useful. It's syntax sucks, though, because it is not a standalone
executable and therefore has to be invoked via kitty.

Note that there exists an `icat` executable in `sleuthkit`, but it is
quite unlikely that I will use it in the foreseeable future, so I'm not
concerned with the name class.