zsh: Try out powerlevel-10k theme
I think this one is as close to perfect as I've seen.

* It can be set to 2-line style which solves problems in narrow
  terminals.
* It can be set to be relatively lean.
* It has git status

I'll try it out for now and see if causes any problems but it's
definitely better than previous themes I've tried. One thing I'll have
to figure out before I can use it widely is disabling the git plugin for
network mounts. EDIT: Or do I? The theme seems to be pretty smart. When
going into a network mount it seemed to disable the git plugin. I don't
know if it has to do with the mount or if it simply realized it was too
slow, but in any case it solves the problem. There's only a minor issue
of it not being re-enabled afterwards but I can live with that.