zsh: Use the ssh kitten to fix ssh problems
Because kitty's terminfo is often missing on the remote, ssh sessions
usually break the terminal. I used to get around this by setting an ssh
alias, but as it turns out, kitty can transfer the terminfo
automatically with this kitten. The benefit compared to the old method
is that no functionality is lost by changing the TERM variable.