Fix 'ssh' alias condition
It's not enough to check that TERM is set to kitty, we also have to make
sure that kitty is installed on the current system. This fixes cases
where TERM is set because we ssh'd in from a kitty system and now we're
trying to ssh again from the host which does not have kitty installed.