Workaround for wofi being stupid with its path resolution
wofi shows all executables on the path, breaking the convention where
binaries can override others of the same name if they appear first.
This breaks stuff that depends on this behavior like firejail's
symlinks.

The solution is not great because it keeps around dmenu which I wanted
tto replace. At least I don't have to deal with its gui which was the
issue but I'd like to get rid of it entirely.