git: Store credentials
This change is made to comply with github's change that requires access
tokens for git command line usage.

Note that the credentials are stored in plaintext which sucks, but
firejail blacklists the fire by default, so any program that runs
through it will not be able to read the credentials.

Also allow pushing into non-bare repositories with this change. This
simplifies some workflows with remotely mounted git repos.