kitty: Use source code pro for italic font
Not sure it's a great fit, but the existence of an italic font at all is
a significant improvement for vim's syntax highlighting. Ideally I would
want Fira Code to have an italic font, but that doesn't seem to exist.
Ideally I would want a font that
 - has pretty characters
 - supports ligatures
 - has an italic variant

So far, the best candidates only have 2/3 of the above. Until I find one
that covers all of it, I'll just use this combo.