Add discord
The line we really care about is

    "SKIP_HOST_UPDATE": true

This should disable the update check when discord starts, which is a
nuisance on package-managed systems. In theory, if discord pushes an
update and arch hasn't shipped it yet, discord will not launch, though I
haven't hit that timing window so far.

The rest of the values are essentially noise, but discord seems to be
one of those apps that abuses its configuration with values that change
frequently and implicitly at runtime. Hopefully these won't change since
I pretty much use discord the same way each time.