wayland: Add xcb as fallback Qt plugin
This should fix applications not starting because they don't support the
wayland plugin. This should be relatively rare. I think it only occurs
for apps that bundle the Qt library, because my system's Qt should work
on Wayland. So this is for apps that release AppImages or similar, like
yuzu.