waybar: Revert hyprland-related changes
I'm going to take a break from hyprland as it seems to still have some
issues to iron out. If/when I return to it I think I'm going to figure
out some setup with separate configs for each window manager instead of
trying to make the single config compatible with both.

This reverts commit 94ca3ca1d61ebce8376de50641a6a5577a0f9982.