nvim: Migrate to astronvim v4
With the astronvim update v3 -> v4, it is now installed and handled as a
"regular" lazy plugin. This greatly simplifies setting it up and
combining it with other plugins and with custom configuration. This is
also a good time to merge my nvim and astronvim configurations.

For now, the only custom configuration is in
`.config/nvim/lua/polish.lua` and the rest of the files are taken from
astronvim's template for installing v4.