zsh: Revert "Auto-launch sway on login"
This reverts commit eec4b4dde02ea9672ba58dfd1d2068201fe22c0c.

I was experimenting with skipping the display manager entirely, but I'm
not entirely satisfied with the results. It causes various minor changes
to things like logging and env vars. I'm not sure I want to commit to a
setup without display manager enough to find and make all the changes I
need. Let's leave it as is for now.