Fix astrolsp configuration being ignored
I noticed that my astrolsp configuration was not working. I'm not sure
exactly what the issue was but I think it had to do with initialization
order. Moving astronvim configuration inside `lazy_setup.lua` seems to
fix it. This does not fix the issue of my clipboard setting in
`polish.lua` being ignored unfortunately.

Additionally I noticed the example configuration has changed by adding
an `enabled` flag inside `format_on_save`. I don't know if this was also
an issue because the above one was breaking the configuration anyway,
but let's fix it anyway.