nvim: Create new simpler config based on LazyVim
First, it's important to note that LazyVim is not the same as lazy.nvim.
LazyVim is an IDE-style setup. lazy.nvim, or commonly just "lazy", is a
package manager for neovim.

The previous setup was based on astronvim. It was great in many regards
but I found it impossible to troubleshoot because it had too many
components I didn't know. Lazyvim is (or seems) smaller and easier to
manage for me. It seems to simply aggregate 3rd party plugins with
custom configuration.

This will probably serve as a stepping stone to purely self-managed
configuration. I don't want this style of all-in-one setups because they
almost always contain stuff I don't like. LazyVim documents its default
plugins [here][1]. I can work through the list to figure out what I
actually want, then create configuration based purely on lazy (the
package manager)

1: https://www.lazyvim.org/plugins