sway: Remove primary monitor
This didn't appear to work reliably. I'm guessing it had to do with
either the monitor or xwayland not being ready when the command ran. If
I reloaded sway's config later manually it would work. Rather than worry
about the timing and trying to find a workaround I just added it to a
wrapper script for Steam and it appears to be much more robust. No
reason to have it here any longer.